Avoid saving preferences when unauthenticated

+++ resources/js/components/Preference.js
@@ -16,10 +16,14 @@ export default class Preference {
     }
 
     set(key, value) {
+        if (!Statamic.$config.get('user')) return;
+
         return this.commitOnSuccessAndReturnPromise(axios.post(this.url, { key, value }));
     }
 
     append(key, value) {
+        if (!Statamic.$config.get('user')) return;
+
         return this.commitOnSuccessAndReturnPromise(axios.post(this.url, { key, value, append: true }));
     }
 
@@ -28,6 +32,8 @@ export default class Preference {
     }
 
     remove(key, value = null, cleanup = true) {
+        if (!Statamic.$config.get('user')) return;
+
         return this.commitOnSuccessAndReturnPromise(axios.delete(`${this.url}/${key}`, { data: { value, cleanup } }));
     }
